vocab/requirements.txt
google-cloud-bigquery>=3.17.0
google-cloud-storage>=2.14.0
